+++ packages/mrm-task-codecov/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.3](https://github.com/sapegin/mrm/compare/mrm-task-codecov@2.1.2...mrm-task-codecov@2.1.3) (2020-11-04)
+
+**Note:** Version bump only for package mrm-task-codecov
+
+
+
+
+
 ## [2.1.2](https://github.com/sapegin/mrm/compare/mrm-task-codecov@2.1.1...mrm-task-codecov@2.1.2) (2020-10-27)
 
 **Note:** Version bump only for package mrm-task-codecov