+++ packages/mrm-task-package/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.5](https://github.com/sapegin/mrm/compare/mrm-task-package@2.1.4...mrm-task-package@2.1.5) (2020-11-04)
+
+**Note:** Version bump only for package mrm-task-package
+
+
+
+
+
 ## [2.1.4](https://github.com/sapegin/mrm/compare/mrm-task-package@2.1.3...mrm-task-package@2.1.4) (2020-10-27)
 
 **Note:** Version bump only for package mrm-task-package